refactor(oauth): separate videos
